fix: Update logo URLs in rspress.config.ts
Currently the logo isn't displayed on any documentation pages (docs subfolder, anything underneath https://extension.js.org/docs/), as the reference is not correctly to the root. Due to different structure, development on localhost, etc. changing it to a different path might not be the best idea.

If you'd like to prevent that absolute path, hosting these two assets in this repository as well instead of crossreferencing them might be an even more resilient approach.